requirements.txt
Flask==2.1.0
keras==2.11.0
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.7.0.72
pandas==1.2.0
Pillow==9.5.0
tensorflow==2.11.0
Gunicorn